abstract Reaction Events
